requirements.txt
mysql_connector_repackaged==0.3.1
plotly==5.2.1
tqdm==4.62.0
Whoosh==2.7.4
